#480-duplicate-requests - delete duplicate requests, update services and component, remove unused service
