+++ src/app/shared/shared.module.ts
@@ -101,7 +101,6 @@ import { ResourceLimitService } from './services/resource-limit.service';
 import { ResourceUsageService } from './services/resource-usage.service';
 import { RouterUtilsService } from './services/router-utils.service';
 import { SecurityGroupService } from './services/security-group.service';
-import { ServiceOfferingFilterService } from './services/service-offering-filter.service';
 import { ServiceOfferingService } from './services/service-offering.service';
 import { SessionStorageService } from './services/session-storage.service';
 import { SnapshotService } from './services/snapshot.service';
@@ -286,7 +285,6 @@ import { ZoneService } from './services/zone.service';
     RouterUtilsService,
     SecurityGroupService,
     ServiceOfferingService,
-    ServiceOfferingFilterService,
     SnapshotService,
     SSHKeyPairService,
     StatsUpdateService,
